fix(connectors): make composed operation loading independent of import order
